test
Research for deeper understanding of the priority NCD of the community, issues with substance abuse and/or mental health needs.
Start building the plan for following weeks based on community and NGO needs, focusing on the issue that is more relevant for the community in the moment.
Run an analysis on the NGO progress and focus regarding NCD, substance abuse and/or mental health initiatives.
Support the staff in carrying out day to day activities
Getting to know the NGO/non-profit representatives.
test

Research for deeper understanding of the priority NCD of the community, issues with substance abuse and/or mental health needs.
Start building the plan for following weeks based on community and NGO needs, focusing on the issue that is more relevant for the community in the moment.
Run an analysis on the NGO progress and focus regarding NCD, substance abuse and/or mental health initiatives.
Support the staff in carrying out day to day activities
Getting to know the NGO/non-profit representatives.test

Research for deeper understanding of the priority NCD of the community, issues with substance abuse and/or mental health needs.
Start building the plan for following weeks based on community and NGO needs, focusing on the issue that is more relevant for the community in the moment.
Run an analysis on the NGO progress and focus regarding NCD, substance abuse and/or mental health initiatives.
Support the staff in carrying out day to day activities
Getting to know the NGO/non-profit representatives.test

Research for deeper understanding of the priority NCD of the community, issues with substance abuse and/or mental health needs.
Start building the plan for following weeks based on community and NGO needs, focusing on the issue that is more relevant for the community in the moment.
Run an analysis on the NGO progress and focus regarding NCD, substance abuse and/or mental health initiatives.
Support the staff in carrying out day to day activities
Getting to know the NGO/non-profit representatives.